useUi -> useSetup
